Insert missing subject.
An alternative way to make the sentence grammatical would be to remove the comma before "but". However, this would have the effect of making "It" (the solution) be the subject of "would love"; as solutions are typically incapable of love, it is dubious that this is the author's original intended meaning. Therefore, the insertion of "I" is the fix that most likely preserves the author's intent.