Insert missing subject.
An alternative way to make the sentence grammatical would be to remove the comma before "but". However, this would have the effect of making "It" (the solution) be the subject of "would love"; as solutions are typically incapable of love, it is dubious that this is the author's original intended meaning. Therefore, the insertion of "I" is the fix that most likely preserves the author's intent.

Former-commit-id: 0cc27366f011121344725ce0109fb3f4be8d76b7
Former-commit-id: 40ded8ecfaf2f3eb881992a6ea1a4c07fdc66305
Former-commit-id: 0350501b6ddb826b70560153a61107233020d62b [formerly e3ca6b8ca9eb3f1eacbb6dad3cd4670c9cf4574f]
Former-commit-id: 3df6ead6bde2083a839a9be73599b1a74bad01b5